TCOMP-838 : migrate salesforce to tacokit
In this version we migrate the input component that relay on salesforce bulk query that are used in web applicaiton to retreive basic datastore and dataset.
The component is covered by unit test that use mocked response.

To run the test in integration mode (directly hit salesforce) you need to define salesforce credentials and activate the passthrough mode